qt-core: qt-cpp: qt-lib: Extract shared StateManager base classes
Move the duplicated Workspace/GlobalStateManager constructors into
BaseWorkspaceStateManager and BaseGlobalStateManager in qt-lib;
qt-core and qt-cpp now keep only their specific state accessors.
