+++ qt-core/src/state.ts
@@ -1,30 +1,15 @@
 // Copyright (C) 2024 The Qt Company Ltd.
 // SPDX-License-Identifier: LicenseRef-Qt-Commercial OR LGPL-3.0-only
 
-import * as vscode from 'vscode';
-
 import {
-  createLogger,
   QtInsRootConfigName,
-  BaseStateManager,
+  BaseWorkspaceStateManager,
+  BaseGlobalStateManager,
   AdditionalQtPathsName,
-  QtAdditionalPath,
-  CoreKey
+  QtAdditionalPath
 } from 'qt-lib';
 
-const logger = createLogger('state');
-
-export class WorkspaceStateManager extends BaseStateManager {
-  constructor(
-    context: vscode.ExtensionContext,
-    folder: vscode.WorkspaceFolder
-  ) {
-    if (folder.uri.fsPath === '') {
-      logger.error('folder is empty');
-      throw new Error('folder is empty');
-    }
-    super(context, folder);
-  }
+export class WorkspaceStateManager extends BaseWorkspaceStateManager {
   public getQtInstallationRoot(): string {
     return this._get<string>(QtInsRootConfigName, '');
   }
@@ -43,10 +28,7 @@ export class WorkspaceStateManager extends BaseStateManager {
   }
 }
 
-export class GlobalStateManager extends BaseStateManager {
-  constructor(context: vscode.ExtensionContext) {
-    super(context, CoreKey.GLOBAL_WORKSPACE);
-  }
+export class GlobalStateManager extends BaseGlobalStateManager {
   public getQtInstallationRoot(): string {
     return this._get<string>(QtInsRootConfigName, '');
   }
